requirements.txt
Flask==3.0.3
flask-cors==4.0.0
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4
scikit-learn==1.5.2
python-dotenv==1.0.1
openai==1.3.0
joblib==1.3.2
gunicorn==21.2.0
pyarrow==15.0.2
tqdm==4.66.4